refactor health schemas
